app/requirements.txt
fastapi[standard]>=0.115,<1.0
asyncpg>=0.30,<0.31
sqlalchemy[asyncio]>=2.0,<3.0
pyjwt-2.10.1>=2.10,<2.11
passlib[bcrypt]>=1.7,<1.8